+++ test/endpoint.js
@@ -36,6 +36,7 @@ describe('classification', function() {
 });
 
 describe('clustering', function() {
+  require('./clustering/inverted-index.js');
   require('./clustering/k-means.js');
   require('./clustering/key-collision.js');
   require('./clustering/naive.js');